Adding a Highscore mode

+++ Source/ArcadeShooter/WaveStruct.h
@@ -39,5 +39,5 @@ struct FWaveStruct : public FTableRowBase
 	int PlanetDestroyerCount;
 
 	UPROPERTY(EditDefaultsOnly)
-	int Time;
+	float Time;
 };
